+++ test/catalog/choice_picker_test.dart
@@ -3,7 +3,7 @@
 // ignore_for_file: missing_whitespace_between_adjacent_strings
 
 import 'package:genui_jaspr/genui_jaspr.dart';
-import 'package:genui_jaspr/src/catalog/components/choice_picker.dart';
+import 'package:genui_jaspr/src/catalog/basic/components/choice_picker.dart';
 import 'package:jaspr_test/jaspr_test.dart';
 
 import '../support/harness.dart';