+++ test/catalog/image_test.dart
@@ -1,6 +1,6 @@
 import 'package:a2ui_core/a2ui_core.dart';
 import 'package:genui_jaspr/genui_jaspr.dart';
-import 'package:genui_jaspr/src/catalog/components/image.dart';
+import 'package:genui_jaspr/src/catalog/basic/components/image.dart';
 import 'package:json_schema_builder/json_schema_builder.dart';
 import 'package:test/test.dart';